requirements.txt
# Esports Computer Vision Project Dependencies
# Optimized for Mac Silicon (M-series/MPS)

# Deep Learning
torch==2.9.1
torchvision==0.24.1

# Object Detection (YOLO)
ultralytics==8.4.7

# Video Processing
yt-dlp==2025.12.8
opencv-python==4.13.0.90

# Data Science
pandas==3.0.0
scikit-learn==1.8.0
xgboost>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0

# OCR
easyocr>=1.7.0

# Utilities
tqdm>=4.66.0
python-pptx>=0.6.23
pyyaml>=6.0
